Move Navbar related components to the same folder
